pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "pymupdf>=1.26.7",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "reportlab>=4.4.6",
]
